build: convert e2e app to standalone
Converts the e2e app to standalone APIs to reduce the amout of code we need to maintain.

(cherry picked from commit 4888ffdbc67317905e23cadf520b69fd47bb37ef)
