+++ src/e2e-app/components/block-scroll-strategy/block-scroll-strategy-e2e.css
