+++ src/e2e-app/components/e2e-app/e2e-app.html
@@ -5,5 +5,5 @@
     <a mat-list-item *ngFor="let link of navLinks" [routerLink]="[link.path]">{{link.title}}</a>
   </mat-nav-list>
 
-  <ng-content></ng-content>
+  <router-outlet></router-outlet>
 </main>
