+++ src/e2e-app/components/home.ts
@@ -1,0 +1,8 @@
+import {Component} from '@angular/core';
+
+@Component({
+  selector: 'home',
+  template: `<p>Welcome to the e2e tests app</p>`,
+  standalone: true,
+})
+export class Home {}
